move logic for flags into separate function
